user can click into routine to find habit, header and sidebar is broken though
