+++ client/components/userRoutine.jsx
@@ -1,47 +1,19 @@
-/*
-* Accessed through sidebar only
-* Send routine to user leads to the routine request page
-*/
-
 import React from 'react';
-import Header from './header';
-import Sidebar from './sidebar';
-import RoutineList from './routineList';
-import Footer from './footer';
-import BlankCard from './blankCard';
+import UserRoutineMain from './userRoutineMain';
+import UserRoutineHabit from './userRoutineHabit';
 
-const UserRoutine = props => {
-  const [routine, setRoutine] = React.useState(null);
-  const [blank, setBlank] = React.useState(false);
-  const userId = 2;
+const UserRoutine = () => {
+  const [view, setView] = React.useState('main');
 
-  function isSideBarOpen() {
-    if (props.isOpen) {
-      return <Sidebar sideRender={'inRoutines'} closeSideBar={props.openSideBar} />;
-    }
-  }
-
-  React.useEffect(
-    () => {
-      fetch(`/api/routine/user/${userId}`)
-        .then(res => res.json())
-        .then(res => setRoutine(res));
-    }, []
-  );
-
-  const createBlank = () => {
-    return blank && <BlankCard setBlank={setBlank}
-      user={userId} routine={routine} setRoutine={setRoutine} />;
+  const createPage = () => {
+    if (view === 'main') return <UserRoutineMain setView={setView} />;
+    else return <UserRoutineHabit routineId={view} />;
   };
 
   return (
-    <div className="bg-light h-100">
-      <Header title={'User Routines'} headerView={'main'} openSideBar={props.openSideBar} />
-      {isSideBarOpen()}
-      <RoutineList routine={routine} userId={userId} />
-      {createBlank()}
-      <Footer setBlank={setBlank} />
-    </div>
+    <>
+      {createPage()}
+    </>
   );
 };
